Separate caching functions and structs to a separate file
